% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkWarehouse.R
\name{perm_test}
\alias{perm_test}
\title{Permutation test for Bulk RNAseq}
\usage{
perm_test(
  DE_list,
  test_genes,
  pval,
  l2fc = 0,
  adjusted = FALSE,
  runs = 10000,
  binwidth = 0.1
)
}
\arguments{
\item{DE_list}{A df of differentially expressed genes from DEseq2 with a pvalue and/or padj and/or log2FoldChange, IMPORTANT: the name of the gene must be in a column titled Gene.name and be in the proper format e.g. Tbr2}

\item{test_genes}{A list of gene names that you wish to test for overrepresentation at the top of your DE gene df}

\item{pval}{A pvalue cutoff for which genes with a smaller pvalue will be allowed to be considered part of the "top DE genes"}

\item{l2fc}{A log2FoldChange cutoff for which genes with a grater absolute value of log2FoldChange will be allowed to be considered part of the "top DE genes"}

\item{adjusted}{A boolean that determines if the pvalue or adjusted pvalue will be considered with the above pval parameter}

\item{runs}{The number of simulations to run}

\item{binwidth}{Binwodth parameter for plotting resulting emperical pvalue histogram}
}
\value{
Empirical pvalue
}
\description{
Generate an empirical p-value to determine if a given gene set is preferentially represented, more frequently than chance, in the top of a differnetilaly expressed gene list
}
\examples{
pvalue <- perm_test(DE_list=DEseq2_results, test_genes=toupper(my_gwas_genes$Gene.name), pval=0.05, runs=10000, l2fc=0.3);
}
