% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkWarehouse.R
\name{readjust_DESeq2_results}
\alias{readjust_DESeq2_results}
\title{Multiple Testing Recorrection of DESeq2 results}
\usage{
readjust_DESeq2_results(res.obj, filt.out.genes.vec)
}
\arguments{
\item{res.obj}{a results dataframe from DEseq2}

\item{filt.out.genes.vec}{A vector of gene ensembl ids to remove from the anlaysis for the purposed of multiple test correction}
}
\value{
A dataframe same as the input dataframe but with two extra columns. pvalue.filt now contains the original pvalue or NA if the gene was filtered out. padj_BH now contains the BenjaminiHochberg adjusted pvalue based on the genes that pass the filter
}
\description{
After identifying genes that should be filtered out of a bulk analysis (e.g. due to low read counts), recorrect the pvalues with multiple testing correction via the Benjamini Hochberg procedure
}
\examples{
recorrected_DESeq2_resutls <- readjust_DESeq2_results(res.obj=my_DESeq2_results, filt.out.genes.vec=c("ENSMUSG00000000766","ENSMUST00000036653.5"));
}
